add single script to flash, run mqtt commands, and write hass config
