+++ README.md
@@ -23,7 +23,7 @@ Create a `config/sites.json` file for your location based on `config/sites.json.
 # To program tasmota devices
 1. Create a device configuration file for your device(s) as in tasmota.json.
     - Each device can be a single device or many identical devices which only differ by some %id% string.
-2. Run flash.py.
+2. Run `./provision.py flash [-m {wifi|serial}] [-p SERIALPORT] config/your_devices.json` (defaults to serial and /dev/ttyUSB0)
 
 `blank_defines.h` undefines all of tasmota's config and lets us set what we want.
 It's processed by python's .format to insert some specific settings near the bottom.
@@ -31,12 +31,12 @@ A version of `platformio_override.ini` is provided here to be used with tasmota.
 
 # To program custom devices
 1. Create a device configuration file for your custom device(s) as in custom.json. Software should be the parent folder. Module should be the name of the program itself. (i.e. you would have your program stored as `./software/module/module.ino`). As above, `%id%` can be used for identical devices.
-2. Run flash.py.
+2. Run `./provision.py flash [-m {wifi|serial}] [-p SERIALPORT] config/your_devices.json` (defaults to serial and /dev/ttyUSB0)
 
 # Other Stuff
-- Query tasmota device status (IP, MAC, tasmota version, core version) with ip_query.py.
-- Write user_config_override.h for a tasmota device with config_tas.py.
-- Home assistant configuration files can be generated via espq.write_hass_config().
+- Query tasmota device status (IP, MAC, tasmota version, esp core version) with ip_query.py.
+- Home assistant configuration files can be generated via `./provision.py hass config/your_devices.json`
+- Run the setup commands from a device's config with `./provision.py cmds config/your_devices.json`
 
 # Device configuration file explanation
 ## name
